Complete Overhaul of Design
New design with a light colored background and dark text. A new
navigation bar on the left side that moves with the user scrolling. Also
added a custom "Fork me on GitHub" link and ribbon. The new design helps
set the stage for any future content or plans I may have, and also
modernizes the look, feel, and functionality of the site.